src/_migrato/_old/requirements.txt
opencv-python>=4.5.0
paho-mqtt>=1.6.0
numpy>=1.20.0
